+++ src/application.lisp
@@ -744,8 +744,7 @@ not supplied. Returns the selected webapp. Convenience function for the REPL."
   "Reset sessions on a per-webapp basis"
   ;; TODO: seems I've removed some code, related to per-webapp sessions
   ;;       probably all code which uses it, should be removed
-  (setf (session-value (webapp-session-key app)) nil)
-  )
+  (setf (session-value (webapp-session-key app)) nil))
 
 ;; Removed in preference to new dependencies
 ;; (defun webapp-application-dependencies (&optional (app *current-webapp*))